Rewrite the bot's update logic
Prior to this on top of having a lot of legacy checks from before
the CI compiled the bot for updating, it also was using a non semvar
compliant check to see if a verison was newer than the bot.

This commit changes that to use the semvar compliant crate semvar to
check the verison to make sure that the bot in the future won't see
a version as newer when it isn't or think one verison is the same as
the one it's running.
(eg 1.0.0 equaling 0.10.0 based on the old logic)