Cleanup admin command
This changes the scopes in the admin command to better limit the amount
of time we hold a lock, additonally, just generally cleanup how it
responds to a user.

This also generally improves the code quality in the command, it's still
not fantastic, but it's far improved from what it was before this
commit.