Add a guild generic reaction roles system with new command to manage it
This commit completely rewrites the previous reaction role system to now
use one backed by a database, and is able to be used across multiple
guilds. It will add and remove roles based on the reactions it receives,
and is far less "jank" than before.

We also add a command to manage this new system called "reaction" which
allows server owners/mods with the MANAGE_ROLES permissions to add and
update (remove will come in another commit) which reactions will give
roles to whatever message they specify.