(feat) Finish Reaction_Add
This completes all the work I had planned for reaction_add.

This commit additionally adds the new dependency Regex, Regex is used
to parase discord message links for the channel and message id inside of
them.
I use an exact match on discordapp.com in order to limit to that domain.

This commit also introduces one new TODO, which involves adding the
lazy_static crate, to allow for reuse of the allocations made by Regex,
for a preformance improvement. Real world this is likely unnecessary for
how often this function will be called, and for how simple the regex is.
It is just a nice to have.