(refactor) Add error handling to the reaction event handlers
Previously, when an error was hit in reaction_add or remove, it would
make an error in the console regardless of whether we cared about the
error or not.

This refactor allows us to more selectivly warn and error on certain
conditions. For example, if the database returns no rows, we don't care
about that, whereas if there's some error from the database, we do care
about what it is.